requirements.txt
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
PyGithub==2.1.1
cryptography==42.0.2
python-dotenv==1.0.1
nbformat==5.9.2
nbconvert==7.14.2
